changed tsconfig.app.json

+++ Controllers/EmployeeItemsController.cs
@@ -1,8 +1,9 @@
+using System.Linq;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-
+using Microsoft.AspNetCore.Authorization;
 
 
 namespace Bif4.SasanJaghori.EmployeeApp
@@ -19,6 +20,7 @@ namespace Bif4.SasanJaghori.EmployeeApp
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> CreateToDoItem([FromBody] EmployeeItem employeeItem)
         {
             if (!ModelState.IsValid)
@@ -28,6 +30,12 @@ namespace Bif4.SasanJaghori.EmployeeApp
 
             employeeItem.Id = 0;
             dbContext.EmployeeItems.Add(employeeItem);
+
+            var claims = User.Claims;
+            employeeItem.UserSubject = claims.FirstOrDefault(c => c.Type == "sub")?.Value;
+            employeeItem.Email = claims.FirstOrDefault(c => c.Type == "email")?.Value;
+            employeeItem.Name = claims.FirstOrDefault(c => c.Type == "name")?.Value;
+
             await dbContext.SaveChangesAsync();
 
             return Ok();
@@ -47,12 +55,19 @@ namespace Bif4.SasanJaghori.EmployeeApp
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> Delete(int id)
         {
             var item = await dbContext.EmployeeItems.FindAsync(id);
             if (item == null)
             {
                 return NotFound();
+            }
+
+            var claims = User.Claims;
+            if(item.UserSubject != claims.FirstOrDefault(c => c.Type == "sub")?.Value)
+            {
+                return Forbid();
             }
 
             dbContext.EmployeeItems.Remove(item);
